% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data-crit.R
\docType{data}
\name{radf_crit}
\alias{radf_crit}
\alias{radf_crit2}
\title{Simulated Monte Carlo critical values}
\format{
A list with lower level lists that contain
\describe{
  \item{adf_cv:}{Augmented Dickey-Fuller}
  \item{badf_cv:}{Backward Augmented Dickey-Fuller}
  \item{sadf_cv:}{Supremum Augmented Dickey-Fuller}
  \item{bsadf_cv:}{Backward Supremum Augmented Dickey-Fuller}
  \item{gsadf_cv:}{Generalized Supremum Augmented Dickey Fuller}

}
}
\source{
simulated from exuber package function \code{\link[=radf_mc_cv]{radf_mc_cv()}}
}
\usage{
radf_crit2
}
\description{
A dataset containing simulated critical values for up to 600 observations
based on default minimum window. The critical values have been simulated and
stored as data to save computation time for the user. The stored critical values
can be obtained with the \code{\link[=radf_mc_cv]{radf_mc_cv()}} function, using 
`nrep = 2000` and `seed = 123`.
}
\examples{
\dontrun{
library(exuber)
all.equal(radf_crit2[[50]], exuber::radf_mc_cv(50, nrep = 2000, seed = 123))
}
}
\keyword{datasets}
